# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VentureForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Ideas
...
  /ideas/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for an idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to an idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (author or admin)
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Contributions
  /ideas/{id}/contributions:
    get:
      operationId: listContributions
      tags: [contributions]
      summary: List contribution requests for an idea (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contribution requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contribution"

    post:
      operationId: offerContribution
      tags: [contributions]
      summary: Offer a contribution to an idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContributionBody"
      responses:
        "201":
          description: Contribution submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contribution"
  /contributions/{contributionId}/status:
    put:
      operationId: updateContributionStatus
      tags: [contributions]
      summary: Accept or reject a contribution (idea owner)
      parameters:
        - name: contributionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContributionStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contribution"

  # User profiles
...
  /admin/ideas:
    get:
      operationId: adminListIdeas
      tags: [admin]
      summary: Admin - list all ideas
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All ideas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedIdeas"
  /admin/ideas/{id}/status:
    put:
      operationId: adminUpdateIdeaStatus
      tags: [admin]
      summary: Admin - approve or remove an idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateIdeaStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /admin/users/{id}/role:
    put:
      operationId: adminUpdateUserRole
      tags: [admin]
      summary: Admin - update user role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRoleBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/comments:
    get:
      operationId: adminListComments
      tags: [admin]
      summary: Admin - list all comments
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedComments"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
      required: [email, password, name]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        skills:
          type: string
          nullable: true
        areasOfInterest:
          type: string
          nullable: true
        willingRoles:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, createdAt]
...
    Idea:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        industry:
          type: string
        targetRegion:
          type: string
        maturityStage:
          type: string
        tags:
          type: string
          nullable: true
        upvotes:
          type: integer
        downvotes:
          type: integer
        score:
          type: integer
        commentsCount:
          type: integer
        contributorsNeeded:
          type: integer
        status:
          type: string
          enum: [active, removed]
        ownerId:
          type: integer
        ownerName:
          type: string
        createdAt:
          type: string
          format: date-time
        userVote:
          type: string
          enum: [up, down]
          nullable: true
        isFollowing:
          type: boolean
          nullable: true
      required: [id, title, summary, industry, targetRegion, maturityStage, upvotes, downvotes, score, commentsCount, contributorsNeeded, status, ownerId, ownerName, createdAt]
...
    PaginatedIdeas:
      type: object
      properties:
        ideas:
          type: array
          items:
            $ref: "#/components/schemas/Idea"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [ideas, total, page, limit, totalPages]
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        ideaId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, ideaId, authorId, authorName, content, createdAt]
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
    Contribution:
      type: object
      properties:
        id:
          type: integer
        ideaId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        role:
          type: string
        message:
          type: string
          nullable: true
        experience:
          type: string
          nullable: true
        contributionType:
          type: string
          enum: [free, equity, payment, investor]
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
      required: [id, ideaId, userId, userName, role, contributionType, status, createdAt]
    CreateContributionBody:
      type: object
      properties:
        role:
          type: string
        message:
          type: string
        experience:
          type: string
        contributionType:
          type: string
          enum: [free, equity, payment, investor]
      required: [role, contributionType]
    UpdateContributionStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, rejected]
      required: [status]
...
    PaginatedUsers:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, limit, totalPages]
    PaginatedComments:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [comments, total, page, totalPages]
    AdminUpdateIdeaStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, removed]
      required: [status]
    AdminUpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
      required: [role]